+++ packages/plugin-rsc/examples/react-router/app/routes/home.tsx
@@ -1,33 +1,19 @@
-namespace Route {
-  export type LoaderArgs = any
-  export type ComponentProps = any
-}
-
 import { sayHello } from './home.actions.ts'
 import { PendingButton } from './home.client.tsx'
 import './home.css'
 import { TestActionStateServer } from './test-action-state/server.tsx'
 
-export function loader({ request }: Route.LoaderArgs) {
-  const url = new URL(request.url)
-  const name = url.searchParams.get('name')
-  return { name: name || 'Unknown' }
-}
-
-const Component = ({ loaderData }: Route.ComponentProps) => {
+const Component = () => {
   return (
     <main className="container my-8 px-8 mx-auto">
       <article className="paper prose max-w-none">
         <h1>Home</h1>
         <p>This is the home page.</p>
         <span className="test-style-home">[test-style-home]</span>
-        <pre>
-          <code>loaderData: {JSON.stringify(loaderData)}</code>
-        </pre>
         <h2>Server Action</h2>
         <form
           className="no-prose grid gap-6"
-          action={sayHello.bind(null, loaderData.name)}
+          action={sayHello.bind(null, 'Demo')}
         >
           <div className="grid gap-1">
             <label className="label" htmlFor="name">
@@ -38,7 +24,7 @@ const Component = ({ loaderData }: Route.ComponentProps) => {
               id="name"
               type="text"
               name="name"
-              placeholder={loaderData.name}
+              placeholder={'Demo'}
             />
           </div>
           <div>