+++ packages/plugin-rsc/types/virtual.d.ts
@@ -1,5 +1,0 @@
-declare module 'virtual:vite-rsc/bootstrap-script-content' {
-  /** @deprecated use `import.meta.viteRsc.loadBootstrapScriptContent("index")` instead */
-  const bootstrapScriptContent: string
-  export default bootstrapScriptContent
-}